fixed some buttons to use POST!

+++ semi_restful_tv_shows/semi_restful_tv_app/views.py
@@ -15,12 +15,13 @@ def newShow(request):
     return render(request,'new_show.html')
 
 def createShow(request):
-    newShow = Show.objects.create(
-        title = request.POST['title'],
-        network = request.POST['network'],
-        release_date = request.POST['release_date'],
-        desc = request.POST['desc']
-    )
+    if request.method == "POST":
+        newShow = Show.objects.create(
+            title = request.POST['title'],
+            network = request.POST['network'],
+            release_date = request.POST['release_date'],
+            desc = request.POST['desc']
+        )
     return redirect('/shows/'+str(newShow.id))
 
 def readShow(request, show_id):
@@ -38,15 +39,17 @@ def editShow(request, show_id):
     return render(request, 'edit_show.html', context)
 
 def updateShow(request, show_id):
-    selected_show = Show.objects.get(id=show_id)
-    selected_show.title = request.POST['title']
-    selected_show.network = request.POST['network']
-    selected_show.release_date = request.POST['release_date']
-    selected_show.desc = request.POST['desc']
-    selected_show.save()
+    if request.method == "POST":
+        selected_show = Show.objects.get(id=show_id)
+        selected_show.title = request.POST['title']
+        selected_show.network = request.POST['network']
+        selected_show.release_date = request.POST['release_date']
+        selected_show.desc = request.POST['desc']
+        selected_show.save()
     return redirect('/shows/'+str(show_id))
 
 def deleteShow(request, show_id):
-    selected_show = Show.objects.get(id=show_id)
-    selected_show.delete()
+    if request.method == "POST":
+        selected_show = Show.objects.get(id=show_id)
+        selected_show.delete()
     return redirect('/')
